performance-measurements/requirements.txt
numpy~=1.26
websocket-client~=1.8
rich~=13.7
